// apps/web/wms-module/types/roles.ts
export enum ROLE_TYPE {
  SUPER_ADMIN = 'super_admin',
  ADMIN = 'admin',
  OPERATOR = 'operator',
  MANAGER = 'manager',
  SANITARIAN = 'sanitarian',
  OPERATOR_TRANSPORTER = 'operator_transporter',
  OPERATOR_LANDFILL = 'operator_landfill',
  OPERATOR_TREATMENT = 'operator_treatment',
  OPERATOR_RECYCLER = 'operator_recycler',
  OPERATOR_SPECIALIZED_TRANSPORT = 'operator_specialized_transport',
  OPERATOR_GOVERMENT = 'operator_government',
  OPERATOR_WASTE_BANK = 'operator_waste_bank',
}

export enum ROLE_LABEL {
  SUPER_ADMIN = 'Super Admin',
  ADMIN = 'Admin',
  FACILITY_ADMIN = 'Facility Admin',
  SANITARIAN = 'Sanitarian',
  MANAGER = 'Manager',
  THIRD_PARTY_ADMIN = 'Third Party Admin',
}